+++ arm9/source/Arm7Patcher.h
@@ -10,6 +10,7 @@ public:
     /// @param loaderPlatform The loader platform to use.
     /// @param cheatsLength The length of the cheats data, or zero when there are no cheats.
     /// @param cheatsPtr Pointer to where the cheats need to be stored, or \c nullptr when there are no cheats.
+    /// @param runInDSiMode \c true if the rom should be loaded in DSi mode, or \c false otherwise.
     /// @return A pointer to the patch space in IWRAM, or \c nullptr if the patches have been placed in main memory.
-    void* ApplyPatches(const LoaderPlatform* loaderPlatform, u32 cheatsLength, void*& cheatsPtr) const;
+    void* ApplyPatches(const LoaderPlatform* loaderPlatform, u32 cheatsLength, void*& cheatsPtr, bool runInDSiMode) const;
 };