+++ arm9/source/Arm9Patcher.h
@@ -23,10 +23,11 @@ public:
     /// @param loaderPlatform The loader platform to use.
     /// @param apListEntry The AP list entry for the rom being loaded, or \c nullptr if there is none.
     /// @param isCloneBootRom \c true if the rom being loaded is a clone boot rom, or \c false otherwise.
+    /// @param runInDSiMode \c true if the rom should be loaded in DSi mode, or \c false otherwise.
     /// @param loaderInfo The loader info to use.
     /// @return Some information resulting from the patching.
     PatchResult ApplyPatches(const LoaderPlatform* loaderPlatform, const ApListEntry* apListEntry,
-        bool isCloneBootRom, const loader_info_t* loaderInfo) const;
+        bool isCloneBootRom, bool runInDSiMode, const loader_info_t* loaderInfo) const;
 
 private:
     const u32* FindMIiUncompressBackward(u32 arm9LoadAddress, SdkVersion sdkVersion) const;