+++ arm9/source/errorDisplay/ErrorDisplay.h
@@ -4,8 +4,21 @@
 class ErrorDisplay
 {
 public:
+    /// @brief Displays the given \p warningString and returns after pressing the A button.
+    /// @param warningString The warning string to display.
+    void PrintWarning(const char* warningString)
+    {
+        Print(warningString, true);
+    }
+
     /// @brief Displays the given \p errorString and loops.
-    /// @note This function does not return.
+    /// @note This function does not return when \p pressAToContinue is \c false.
     /// @param errorString The error string to display.
-    void PrintError(const char* errorString);
+    void PrintError(const char* errorString)
+    {
+        Print(errorString, false);
+    }
+
+private:
+    void Print(const char* errorString, bool pressAToContinue);
 };