+++ arm9/source/patches/arm9/OSResetSystemPatch.h
@@ -6,8 +6,8 @@
 class OSResetSystemPatch : public Patch
 {
 public:
-    explicit OSResetSystemPatch(const loader_info_t* loaderInfo)
-        : _loaderInfo(loaderInfo) { }
+    OSResetSystemPatch(const loader_info_t* loaderInfo, bool runInDSiMode)
+        : _loaderInfo(loaderInfo), _runInDSiMode(runInDSiMode) { }
 
     bool FindPatchTarget(PatchContext& patchContext) override;
     void ApplyPatch(PatchContext& patchContext) override;
@@ -19,7 +19,8 @@ public:
 
 private:
     u32* _osResetSystem = nullptr;
-    u32 _hybrid = false;
+    u16 _hybrid = false;
+    u16 _runInDSiMode;
     const loader_info_t* _loaderInfo;
     void** _cheatsPointer = nullptr;
 };